+++ R/print.effectsize_table.R
@@ -123,10 +123,11 @@ print.equivalence_test_effectsize <- function(x, digits = 2, ...) {
 
 #' @export
 #' @rdname print.effectsize_table
-#' @param append_CL Should the Common Language Effect Sizes be printed as well?
+#' @param append_CLES Should the Common Language Effect Sizes be printed as well?
 #'   Only applicable to Cohen's *d*, Hedges' *g* for independent samples of
-#'   equal variance (pooled sd) (See [d_to_common_language()])
-print.effectsize_difference <- function(x, digits = 2, append_CL = FALSE, ...) {
+#'   equal variance (pooled sd) or for the rank-biserial correlation for
+#'   independent samples (See [d_to_cles()])
+print.effectsize_difference <- function(x, digits = 2, append_CLES = FALSE, ...) {
   x_orig <- x
 
   footer <- caption <- subtitle <- NULL
@@ -153,19 +154,18 @@ print.effectsize_difference <- function(x, digits = 2, append_CL = FALSE, ...) {
   print.effectsize_table(x, digits = digits, ...)
 
 
-  if (append_CL) {
-    if (any(colnames(x_orig) %in% c("Cohens_d", "Hedges_g")) &&
-      attr(x_orig, "pooled_sd") &&
-      !attr(x_orig, "paired")) {
-      # Common lang
-      cl <- d_to_common_language(x_orig[[any(colnames(x_orig) %in% c("Cohens_d", "Hedges_g"))]])
-      cl <- lapply(cl, insight::format_value, as_percent = TRUE, digits = digits)
-      cl <- data.frame(cl, check.names = FALSE)
-      cat(insight::export_table(cl,
-        digits = digits,
-        caption = c("\n\n# Common Language Effect Sizes", "blue"), ...
-      ))
+  if (append_CLES) {
+    if ("r_rank_biserial" %in% colnames(x_orig)) {
+      to_cl_coverter <- rb_to_cles
+    } else {
+      to_cl_coverter <- d_to_cles
     }
+
+    tryCatch({
+      CL <- to_cl_coverter(x_orig)
+      attr(CL, "table_caption") <- c("\n\n# Common Language Effect Sizes", "blue")
+      print(CL, digits = digits)
+    }, error = function(...) invisible(NULL))
   }
 
   invisible(x_orig)